refactor : step2 LottoException 클래스 LottoValidationUtils로 클래스명 변경
