+++ src/main/java/lotto/domain/LottoService.java
@@ -1,13 +1,13 @@
 package lotto.domain;
 
-import lotto.exception.LottoException;
+import lotto.utils.LottoValidationUtils;
 
 public class LottoService {
 
     public static final int LOTTO_TICKET_PRICE = 1000;
 
     public int buyLottoTicket(int price) {
-        LottoException.invalidLottoTicketInputPrice(price, LOTTO_TICKET_PRICE);
+        LottoValidationUtils.invalidLottoTicketInputPrice(price, LOTTO_TICKET_PRICE);
         return price / LOTTO_TICKET_PRICE;
     }
