+++ src/test/java/lotto/domain/LottoNumberGeneratorTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static lotto.exception.LottoException.INVALID_LOTTO_NUMBER;
+import static lotto.utils.LottoValidationUtils.INVALID_LOTTO_NUMBER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
