+++ src/test/java/lotto/domain/LottoServiceTest.java
@@ -1,13 +1,13 @@
 package lotto.domain;
 
-import lotto.exception.LottoException;
+import lotto.utils.LottoValidationUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import static lotto.exception.LottoException.PRINT_INVALID_LOTTO_PRICE;
+import static lotto.utils.LottoValidationUtils.PRINT_INVALID_LOTTO_PRICE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
@@ -33,7 +33,7 @@ class LottoServiceTest {
     void inputPriceValidation() {
         assertThatExceptionOfType(IllegalArgumentException.class)
                 .isThrownBy(() -> {
-                    LottoException.invalidLottoTicketInputPrice(11500, LottoService.LOTTO_TICKET_PRICE);
+                    LottoValidationUtils.invalidLottoTicketInputPrice(11500, LottoService.LOTTO_TICKET_PRICE);
                 }).withMessageMatching(PRINT_INVALID_LOTTO_PRICE);
     }
